+++ libs/megaparse/src/megaparse/formatter/table_formatter/vision_table_formatter.py
@@ -6,10 +6,9 @@ from typing import List, Optional
 from langchain_core.language_models.chat_models import BaseChatModel
 from langchain_core.messages import HumanMessage
 from megaparse.formatter.table_formatter import TableFormatter
-from megaparse.models.document import Document, TableBlock
+from megaparse_sdk.schema.document import Document, TableBlock
 from pdf2image import convert_from_path
 from PIL import Image
-from unstructured.documents.elements import Element
 
 TABLE_OCR_PROMPT = """
 You are tasked with transcribing the content of a table into markdown format. Your goal is to create a well-structured, readable markdown table that accurately represents the original content while adding appropriate formatting.
