+++ libs/megaparse/src/megaparse/layout_detection/output.py
@@ -1,0 +1,11 @@
+from uuid import UUID
+
+from megaparse_sdk.schema.document import BBOX
+from pydantic import BaseModel
+
+
+class LayoutDetectionOutput(BaseModel):
+    bbox_id: UUID
+    bbox: BBOX
+    prob: float
+    label: int
