+++ libs/megaparse/src/megaparse/parser/base.py
@@ -1,11 +1,9 @@
 from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import IO, List
+from typing import IO
 
-from megaparse.models.page import Page
+from megaparse_sdk.schema.document import Document
 from megaparse_sdk.schema.extensions import FileExtension
-
-from megaparse.models.document import Document
 
 
 class BaseParser(ABC):
