+++ libs/megaparse/src/megaparse/utils/extract_metadata.py
@@ -1,0 +1,7 @@
+from typing import Any, Dict
+
+import pypdfium2 as pdfium
+
+
+def get_doc_metdata(pdfium_document: pdfium.PdfDocument) -> Dict[str, Any]:
+    pass
